+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/chunks/CjSEqTGI.js
@@ -1,0 +1,1 @@
+import"./Cd8kmZxy.js";import{i as e}from"./B_1RL_pw.js";var t=()=>{let t=e;return{page:{subscribe:t.page.subscribe},navigating:{subscribe:t.navigating.subscribe},updated:t.updated}},n={subscribe(e){return t().page.subscribe(e)}};export{n as t};